+++ src/Substitutions.hs
@@ -236,12 +236,13 @@ unicodeSubst =
   , ("Ô",           "O")        -- U+00D4 LATIN CAPITAL LETTER O WITH CIRCUMFLEX
   , ("Õ",           "O")        -- U+00D5 LATIN CAPITAL LETTER O WITH TILDE
   , ("Ö",           "O")        -- U+00D6 LATIN CAPITAL LETTER O WITH DIAERESIS
-  , ("Ú",           "U")        -- U+00DA LATIN CAPITAL LETTER U WITH ACUTE
-  , ("Û",           "U")        -- U+00DB LATIN CAPITAL LETTER U WITH CIRCUMFLEX
-  , ("Ü",           "U")        -- U+00DC LATIN CAPITAL LETTER U WITH DIAERESIS
   , ("×",           "")         -- U+00D7 MULTIPLICATION SIGN
   , ("Ø",           "O")        -- U+00D8 LATIN CAPITAL LETTER O WITH STROKE
   , ("Ù",           "U")        -- U+00D9 LATIN CAPITAL LETTER U WITH GRAVE
+  , ("Ú",           "U")        -- U+00DA LATIN CAPITAL LETTER U WITH ACUTE
+  , ("Û",           "U")        -- U+00DB LATIN CAPITAL LETTER U WITH CIRCUMFLEX
+  , ("Ü",           "U")        -- U+00DC LATIN CAPITAL LETTER U WITH DIAERESIS
+  , ("Ý",           "Y")        -- U+00DD LATIN CAPITAL LETTER Y WITH ACUTE
   , ("ß",           "ss")       -- U+00DF LATIN SMALL LETTER SHARP S
   , ("à",           "a")        -- U+00E0 LATIN SMALL LETTER A WITH GRAVE
   , ("á",           "a")        -- U+00E1 LATIN SMALL LETTER A WITH ACUTE
@@ -269,6 +270,7 @@ unicodeSubst =
   , ("ú",           "u")        -- U+00FA LATIN SMALL LETTER U WITH ACUTE
   , ("û",           "u")        -- U+00FB LATIN SMALL LETTER U WITH CIRCUMFLEX
   , ("ü",           "u")        -- U+00FC LATIN SMALL LETTER U WITH DIAERESIS
+  , ("ý",           "y")        -- U+00FD LATIN SMALL LETTER Y WITH ACUTE
   , ("þ",           "t")        -- U+00FE LATIN SMALL LETTER THORN
   , ("ÿ",           "y")        -- U+00FF LATIN SMALL LETTER Y WITH DIAERESIS
   , ("Ă",           "A")        -- U+0102 LATIN CAPITAL LETTER A WITH BREVE
@@ -277,9 +279,14 @@ unicodeSubst =
   , ("ą",           "a")        -- U+0105 LATIN SMALL LETTER A WITH OGONEK
   , ("Ć",           "c")        -- U+0106 LATIN CAPITAL LETTER C WITH ACUTE
   , ("ć",           "c")        -- U+0107 LATIN SMALL LETTER C WITH ACUTE
+  , ("Č",           "C")        -- U+010C LATIN CAPITAL LETTER C WITH CARON
   , ("č",           "c")        -- U+010D LATIN SMALL LETTER C WITH CARON
+  , ("Ď",           "D")        -- U+010E LATIN CAPITAL LETTER D WITH CARON
+  , ("ď",           "d")        -- U+010F LATIN SMALL LETTER D WITH CARON
   , ("Ę",           "e")        -- U+0118 LATIN CAPITAL LETTER E WITH OGONEK
   , ("ę",           "e")        -- U+0119 LATIN SMALL LETTER E WITH OGONEK
+  , ("Ě",           "E")        -- U+011A LATIN CAPITAL LETTER E WITH CARON
+  , ("ě",           "e")        -- U+011B LATIN SMALL LETTER E WITH CARON
   , ("Ğ",           "G")        -- U+011E LATIN CAPITAL LETTER G WITH BREVE
   , ("ğ",           "g")        -- U+011F LATIN SMALL LETTER G WITH BREVE
   , ("Ī",           "I")        -- U+012A LATIN CAPITAL LETTER I WITH MACRON
@@ -290,10 +297,13 @@ unicodeSubst =
   , ("Ń",           "N")        -- U+0143 LATIN CAPITAL LETTER N WITH ACUTE
   , ("ń",           "n")        -- U+0144 LATIN SMALL LETTER N WITH ACUTE
   , ("ņ",           "n")        -- U+0146 LATIN SMALL LETTER N WITH CEDILLA
+  , ("Ň",           "N")        -- U+0147 LATIN CAPITAL LETTER N WITH CARON
+  , ("ň",           "n")        -- U+0148 LATIN SMALL LETTER N WITH CARON
   , ("Ő",           "O")        -- U+0150 LATIN CAPITAL LETTER O WITH DOUBLE ACUTE
   , ("ő",           "o")        -- U+0151 LATIN SMALL LETTER O WITH DOUBLE ACUTE
   , ("Œ",           "OE")       -- U+0152 LATIN CAPITAL LIGATURE OE
   , ("œ",           "oe")       -- U+0153 LATIN SMALL LIGATURE OE
+  , ("Ř",           "R")        -- U+0158 LATIN CAPITAL LETTER R WITH CARON
   , ("ř",           "r")        -- U+0159 LATIN SMALL LETTER R WITH CARON
   , ("Ś",           "S")        -- U+015A LATIN CAPITAL LETTER S WITH ACUTE
   , ("ś",           "s")        -- U+015B LATIN SMALL LETTER S WITH ACUTE
@@ -301,7 +311,11 @@ unicodeSubst =
   , ("ş",           "s")        -- U+015F LATIN SMALL LETTER S WITH CEDILLA
   , ("Š",           "S")        -- U+0160 LATIN CAPITAL LETTER S WITH CARON
   , ("š",           "s")        -- U+0161 LATIN SMALL LETTER S WITH CARON
+  , ("Ť",           "T")        -- U+0164 LATIN CAPITAL LETTER T WITH CARON
+  , ("ť",           "t")        -- U+0165 LATIN SMALL LETTER T WITH CARON
   , ("ū",           "u")        -- U+016B LATIN SMALL LETTER U WITH MACRON
+  , ("Ů",           "U")        -- U+016E LATIN CAPITAL LETTER U WITH RING ABOVE
+  , ("ů",           "u")        -- U+016F LATIN SMALL LETTER U WITH RING ABOVE
   , ("Ű",           "U")        -- U+0170 LATIN CAPITAL LETTER U WITH DOUBLE ACUTE
   , ("ű",           "u")        -- U+0171 LATIN SMALL LETTER U WITH DOUBLE ACUTE
   , ("Ÿ",           "Y")        -- U+0178 LATIN CAPITAL LETTER Y WITH DIAERESIS